Cleanup

+++ platformio/include/client_utils.h
@@ -24,7 +24,7 @@
 #include <PubSubClient.h>
 #endif
 
-wl_status_t startWiFi(int &wifiRSSI);
+wl_status_t startWiFi(int8_t &wifiRSSI);
 void killWiFi();
 bool waitForSNTPSync(tm *timeInfo);
 bool printLocalTime(tm *timeInfo);
